+++ modules/gh-oidc/versions.tf
@@ -20,20 +20,20 @@ terraform {
 
     google = {
       source  = "hashicorp/google"
-      version = ">= 3.64, < 6"
+      version = ">= 3.64, < 7"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 3.64, < 6"
+      version = ">= 3.64, < 7"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-github-actions-runners:gh-oidc/v3.1.2"
+    module_name = "blueprints/terraform/terraform-google-github-actions-runners:gh-oidc/v4.0.0"
   }
 
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-github-actions-runners:gh-oidc/v3.1.2"
+    module_name = "blueprints/terraform/terraform-google-github-actions-runners:gh-oidc/v4.0.0"
   }
 
 }