+++ test/setup/versions.tf
@@ -20,11 +20,11 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 3.53.0, < 6"
+      version = ">= 3.53.0, < 7"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 3.53.0, < 6"
+      version = ">= 3.53.0, < 7"
     }
   }
 }